fix: integer size

+++ main.c
@@ -5,7 +5,7 @@
 #include <windows.h>
 
 // slow function to check prime numbers
-static int isPrime(unsigned long n) {
+static int isPrime(unsigned long long n) {
   if (n <= 1) return 0;
   if (n == 2) return 1;
   if (n % 2 == 0) return 0;
@@ -19,8 +19,8 @@ static int isPrime(unsigned long n) {
 }
 
 // add all prime numbers between the given limits
-static unsigned long addPrimes(unsigned long start, unsigned long end) {
-  unsigned long sum = 0;
+static unsigned long long addPrimes(unsigned long long start, unsigned long long end) {
+  unsigned long long sum = 0;
   for (; start < end; ++start) {
     if (isPrime(start)) sum += start;
   }
@@ -29,32 +29,32 @@ static unsigned long addPrimes(unsigned long start, unsigned long end) {
 }
 
 typedef struct TData {
-  unsigned long start, end;
-  unsigned long sum;
+  unsigned long long start, end;
+  unsigned long long sum;
 } TDATA, *PTDATA;
 
-static void calculateSimple(unsigned long limit) {
-  const unsigned long sum = addPrimes(0, limit);
-  printf("calculate simple (%d): %d\n", limit, sum);
+static void calculateSimple(unsigned long long limit) {
+  const unsigned long long sum = addPrimes(0, limit);
+  printf("calculate simple (%lld): %lld\n", limit, sum);
 }
 
 DWORD WINAPI threadAddPrimes(PTDATA data) {
-  const unsigned long sum = addPrimes(data->start, data->end);
+  const unsigned long long sum = addPrimes(data->start, data->end);
   data->sum = sum;
   return 0;
 }
 
-static void calculateThreaded(unsigned long limit, unsigned tCount) {
+static void calculateThreaded(unsigned long long limit, unsigned tCount) {
   PTDATA pDataArray = calloc(tCount, sizeof(TDATA));
   PHANDLE hThreadArray = calloc(tCount, sizeof(HANDLE));
 
   if (!pDataArray || !hThreadArray) ExitProcess(2);
   
-  unsigned long start = 0;
-  unsigned long step = limit / tCount;
+  unsigned long long start = 0;
+  unsigned long long step = limit / tCount;
   for (unsigned i = 0; i < tCount; ++i) {
     pDataArray[i].start = start;
-    unsigned long end = start + step;
+    unsigned long long end = start + step;
     if (i + 1 == tCount) end = limit;
     pDataArray[i].end = end;
     start = end;
@@ -72,12 +72,12 @@ static void calculateThreaded(unsigned long limit, unsigned tCount) {
 
   WaitForMultipleObjects(tCount, hThreadArray, TRUE, INFINITE);
 
-  unsigned long total = 0;
+  unsigned long long total = 0;
   for (unsigned i = 0; i < tCount; ++i) {
     total += pDataArray[i].sum;
   }
 
-  printf("calculate threaded (%dt, %d): %d\n", tCount, limit, total);
+  printf("calculate threaded (%dt, %lld): %lld\n", tCount, limit, total);
 
   for (unsigned i = 0; i < tCount; ++i) {
     CloseHandle(hThreadArray[i]);
